fix(slack): make the slack form accessible
A visible label has been added to the input to indicate what data the input actually takes. This is useful for both the eye, and particularly the screenreader.

The ‘autofocus’ attribute has been removed from the email input, jumping to a box with no context (i.e the screenreader didn’t get a chance to read the contextual information about IRC/Slack) is not useful.

A better message for when the individual is already a member of the team has been added.

Styling has been added to the messaging on success/fail, with green/red success/fail states.

The button `input` has been changed to `button` as HTML5 is awesome.

The JavaScript was also tidied a little to make it more readable.